Add --latest-group-version flag and improve error handling
Added a new flag --latest-group-version to the channel-make-version script to facilitate the creation of a version based on the very latest version in a group. With this flag, users now have the flexibility to base their version on either a specified version id via --group-version or the most recent version with --latest-group-version. Additionally, to avoid confusion and maintain code preciseness, an error will be raised if neither of these flags is used to specify the group version to import from. This commit includes a version upgrade to 3.3.0